+++ CHANGELOG.md
@@ -71,7 +71,9 @@
 
 #### Enhancements
 
-* None.
+* Enable caching for collecting rules.  
+  [Andrés Cecilia Luque](https://github.com/acecilia)
+  [#3796](https://github.com/realm/SwiftLint/issues/3796)
 
 #### Bug Fixes
 
@@ -232,6 +234,7 @@
 * Fix the cache path not being properly set when using nested
   configurations.  
   [Andrés Cecilia Luque](https://github.com/acecilia)
+  [#3797](https://github.com/realm/SwiftLint/issues/3797)
 
 ## 0.45.1: Clothes Drying Hooks